requirements.txt
maxminddb 
maxminddb-geolite2
matplotlib
annoy
basemap